reorganize verifier
