+++ math/src/polynomial/multilinear_poly.rs
@@ -3,7 +3,6 @@ use crate::field::traits::{IsField, IsPrimeField};
 use crate::polynomial::multilinear_term::MultiLinearMonomial;
 use crate::polynomial::term::Term;
 use core::ops::AddAssign;
-use serde::Serialize;
 use std::fmt::Display;
 
 /// Represents a multilinear polynomials as a collection of multilinear monomials
