+++ math/src/polynomial/multilinear_term.rs
@@ -1,14 +1,13 @@
 use crate::field::element::FieldElement;
 use crate::field::traits::{IsField, IsPrimeField};
 use crate::polynomial::term::Term;
-use serde::Serialize;
 use std::cmp::Eq;
 use std::fmt::Display;
 
 /// Struct for (coeff: FieldElement<F>, terms: Vec<usize>) representing a multilinear
 /// monomial in a sparse format.
 // TODO: add check that var labels are 0 indexed
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MultiLinearMonomial<F: IsPrimeField>
 where
     <F as IsField>::BaseType: Send + Sync,
